build: Fix infinite loop with Makefile Tools VS Code extension
The issue was caused by the `--always-make` flag that the extension always uses for the configure step.
Not automatically printing the manual is actually better because most of the times the user knows what they want to type but just had a typo. A short error message is better here as it shows them what they typed.